Update pre-commit to avoid error messages on pull
Previously, when pulling new changes from remove, an error would appear due to a head mismatch, from the local repo being behind the remote. This change makes it so that no longer happens, but new changes are still brought in. It does this using Merge. 

Please feel free to improve this code. I am no pro. I just used Google Bard to help get rid of the annoying error messages. Thank you for the resources to integrate Git and Logseq, I love it!